Работа с Moxy | Работа с MainActivity.kt.
