+++ app/src/main/res/layout/activity_main.xml
@@ -1,50 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="center_horizontal"
-    android:orientation="vertical"
-    android:padding="@dimen/activity_main_container_paddings">
+    android:layout_height="match_parent">
 
-    <TextView
-        android:id="@+id/main_activity_counter_one_textview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/activity_main_counter_one_textview_text" />
-
-    <Button
-        android:id="@+id/main_activity_counter_one_btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/activity_main_counter_btn_init_text"
-        app:layout_constraintTop_toBottomOf="@id/main_activity_counter_one_textview" />
-
-    <TextView
-        android:id="@+id/main_activity_counter_two_textview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/activity_main_counter_two_textview_text"
-        app:layout_constraintTop_toBottomOf="@id/main_activity_counter_one_btn" />
-
-    <Button
-        android:id="@+id/main_activity_counter_two_btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/activity_main_counter_btn_init_text"
-        app:layout_constraintTop_toBottomOf="@id/main_activity_counter_two_textview" />
-
-    <TextView
-        android:id="@+id/main_activity_counter_three_textview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/activity_main_counter_three_textview_text"
-        app:layout_constraintTop_toBottomOf="@id/main_activity_counter_two_btn" />
-
-    <Button
-        android:id="@+id/main_activity_counter_three_btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/activity_main_counter_btn_init_text"
-        app:layout_constraintTop_toBottomOf="@id/main_activity_counter_three_textview" />
-</LinearLayout>
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/main_activity_users_recycler_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:clipToPadding="false"
+        android:paddingTop="16dp"
+        android:paddingBottom="16dp" />
+</RelativeLayout>
